Added info button about database; fixed minor problems with main GUI.
