+++ scripts/time_stress.py
