+++ tests/datahandling/test_read_sql.py
@@ -1,0 +1,168 @@
+import io
+import json
+import os
+import sqlite3
+import sys
+import tempfile
+import unittest
+from contextlib import redirect_stdout
+from pathlib import Path
+
+os.environ.setdefault("QT_QPA_PLATFORM", "offscreen")
+
+import numpy as np
+from PyQt5 import QtCore
+from PyQt5 import QtGui
+from PyQt5 import QtWidgets as qtw
+import pyqtgraph as pg
+
+from qplot.configuration.config import config
+from qplot.configuration.scripts import scripts, sysHandle, try_as_num
+from qplot.configuration.themes import dark, light
+from qplot.windows import main as main_window
+from qplot.windows import _plotWin as plotwin_module
+from qplot.windows.plot1d import plot1d
+from qplot.windows.plot2d import _COLORBAR_COLORMAPS, plot2d
+from qplot.windows._subplots import custom_viewbox
+from qplot.windows._plotWin import plotWidget
+from qplot.windows._window_controls import (
+    add_confirmation_options,
+    add_restore_defaults_option,
+    )
+from qplot.windows._dragdrop import (
+    make_run_preview_mime,
+    preview_drop_is_compatible,
+    run_preview_payload_from_mime,
+    )
+from qplot.windows._widgets.operations import operations_options_1d
+from qplot.windows._widgets.toolbar import QDock_context
+from qplot.windows._widgets import treeWidgets
+from qplot.windows._widgets.preview import (
+    DraggablePreviewImageLabel,
+    PREVIEW_BACKGROUND_COLOR,
+    PREVIEW_SIZE,
+    PREVIEW_SELECTED_PROPERTY,
+    PreviewTab,
+    generate_run_previews,
+    render_heatmap_preview,
+    render_sparkline_preview,
+    )
+from qplot.datahandling import readSQL
+from qplot.tools.general import data2matrix
+from qplot.tools.plot_tools import differentiate, pass_filter, subtract_mean
+
+
+class RunSizeTestCase(unittest.TestCase):
+    def test_point_shape_uses_largest_measured_parameter_shape(self):
+        self.assertEqual(
+            readSQL._point_shape(
+                {
+                    "shapes": {
+                        "dmm_v1": [10, 100],
+                        "dmm_v2": [10],
+                        }
+                    },
+                ["dmm_v1", "dmm_v2"]
+                ),
+            [10, 100]
+            )
+
+    def test_point_shape_falls_back_to_distinct_sweep_values(self):
+        conn = sqlite3.connect(":memory:")
+        try:
+            cursor = conn.cursor()
+            cursor.execute("CREATE TABLE results (x REAL, y REAL, signal_a REAL, signal_b REAL)")
+            cursor.executemany(
+                "INSERT INTO results VALUES (?, ?, ?, ?)",
+                [
+                    (0.0, 0.0, 1.0, 2.0),
+                    (0.0, 1.0, 3.0, 4.0),
+                    (1.0, 0.0, 5.0, 6.0),
+                    (1.0, 1.0, 7.0, 8.0),
+                    ]
+                )
+
+            self.assertEqual(
+                readSQL._point_shape_from_result_table(
+                    cursor,
+                    "results",
+                    ["x", "y"],
+                    ["signal_a", "signal_b"],
+                    4,
+                    ),
+                [2, 2]
+                )
+        finally:
+            conn.close()
+
+    def test_point_shape_fallback_includes_measured_row_factor(self):
+        conn = sqlite3.connect(":memory:")
+        try:
+            cursor = conn.cursor()
+            cursor.execute("CREATE TABLE results (x REAL, y REAL, signal_a REAL, signal_b REAL)")
+            cursor.executemany(
+                "INSERT INTO results VALUES (?, ?, ?, ?)",
+                [
+                    (0.0, 0.0, 1.0, None),
+                    (0.0, 0.0, None, 2.0),
+                    (0.0, 1.0, 3.0, None),
+                    (0.0, 1.0, None, 4.0),
+                    (1.0, 0.0, 5.0, None),
+                    (1.0, 0.0, None, 6.0),
+                    (1.0, 1.0, 7.0, None),
+                    (1.0, 1.0, None, 8.0),
+                    ]
+                )
+
+            self.assertEqual(
+                readSQL._point_shape_from_result_table(
+                    cursor,
+                    "results",
+                    ["x", "y"],
+                    ["signal_a", "signal_b"],
+                    8,
+                    ),
+                [2, 2, 2]
+                )
+            self.assertEqual(
+                readSQL._setpoint_shape_from_result_table(
+                    cursor,
+                    "results",
+                    ["x", "y"],
+                    ),
+                [2, 2]
+                )
+        finally:
+            conn.close()
+
+    def test_storage_size_falls_back_to_schema_estimate(self):
+        with tempfile.TemporaryDirectory() as temp_dir:
+            database_path = os.path.join(temp_dir, "storage.db")
+            conn = sqlite3.connect(database_path)
+            try:
+                cursor = conn.cursor()
+                cursor.execute("""
+                  CREATE TABLE "results-1-4" (
+                      id INTEGER,
+                      timestamp REAL,
+                      dac_ch1 REAL,
+                      dac_ch2 REAL,
+                      dmm_v1 REAL,
+                      dmm_v2 REAL
+                  )
+                """)
+                cursor.executemany(
+                    'INSERT INTO "results-1-4" VALUES (?, ?, ?, ?, ?, ?)',
+                    [(i, i * 0.01, 1.0, 2.0, 3.0, 4.0) for i in range(2000)]
+                    )
+                conn.commit()
+
+                self.assertEqual(
+                    readSQL._estimated_table_storage_bytes(cursor, "results-1-4"),
+                    112000
+                    )
+            finally:
+                cursor.close()
+                conn.close()
+
+
