+++ tests/windows/test_main_window.py
@@ -1,0 +1,331 @@
+import io
+import json
+import os
+import sqlite3
+import sys
+import tempfile
+import unittest
+from contextlib import redirect_stdout
+from pathlib import Path
+
+os.environ.setdefault("QT_QPA_PLATFORM", "offscreen")
+
+import numpy as np
+from PyQt5 import QtCore
+from PyQt5 import QtGui
+from PyQt5 import QtWidgets as qtw
+import pyqtgraph as pg
+
+from qplot.configuration.config import config
+from qplot.configuration.scripts import scripts, sysHandle, try_as_num
+from qplot.configuration.themes import dark, light
+from qplot.windows import main as main_window
+from qplot.windows import _plotWin as plotwin_module
+from qplot.windows.plot1d import plot1d
+from qplot.windows.plot2d import _COLORBAR_COLORMAPS, plot2d
+from qplot.windows._subplots import custom_viewbox
+from qplot.windows._plotWin import plotWidget
+from qplot.windows._window_controls import (
+    add_confirmation_options,
+    add_restore_defaults_option,
+    )
+from qplot.windows._dragdrop import (
+    make_run_preview_mime,
+    preview_drop_is_compatible,
+    run_preview_payload_from_mime,
+    )
+from qplot.windows._widgets.operations import operations_options_1d
+from qplot.windows._widgets.toolbar import QDock_context
+from qplot.windows._widgets import treeWidgets
+from qplot.windows._widgets.preview import (
+    DraggablePreviewImageLabel,
+    PREVIEW_BACKGROUND_COLOR,
+    PREVIEW_SIZE,
+    PREVIEW_SELECTED_PROPERTY,
+    PreviewTab,
+    generate_run_previews,
+    render_heatmap_preview,
+    render_sparkline_preview,
+    )
+from qplot.datahandling import readSQL
+from qplot.tools.general import data2matrix
+from qplot.tools.plot_tools import differentiate, pass_filter, subtract_mean
+
+
+class CloseAllPlotsTestCase(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.app = qtw.QApplication.instance() or qtw.QApplication([])
+
+    def test_close_all_can_be_cancelled_when_warning_enabled(self):
+        old_question = qtw.QMessageBox.question
+        closed = []
+
+        class FakeConfig:
+            def get(self, key):
+                if key == "user_preference.confirm_close_all":
+                    return True
+                raise KeyError(key)
+
+        class FakeWindow:
+            def close(self):
+                closed.append(self)
+
+        class Harness:
+            closeAll = main_window.MainWindow.closeAll
+
+            def __init__(self):
+                self.config = FakeConfig()
+                self.windows = [FakeWindow()]
+                self.status_messages = []
+
+            def show_status(self, message, timeout=5000):
+                self.status_messages.append((message, timeout))
+
+        try:
+            qtw.QMessageBox.question = lambda *args, **kwargs: qtw.QMessageBox.No
+            harness = Harness()
+            harness.closeAll()
+        finally:
+            qtw.QMessageBox.question = old_question
+
+        self.assertEqual(closed, [])
+        self.assertEqual(harness.status_messages[-1][0], "Close all plot windows cancelled.")
+
+    def test_close_all_without_warning_closes_each_window(self):
+        closed = []
+
+        class FakeConfig:
+            def get(self, key):
+                if key == "user_preference.confirm_close_all":
+                    return False
+                raise KeyError(key)
+
+        class FakeWindow:
+            def close(self):
+                closed.append(self)
+
+        class Harness:
+            closeAll = main_window.MainWindow.closeAll
+
+            def __init__(self):
+                self.config = FakeConfig()
+                self.windows = [FakeWindow(), FakeWindow()]
+                self.status_messages = []
+
+            def show_status(self, message, timeout=5000):
+                self.status_messages.append((message, timeout))
+
+        harness = Harness()
+        harness.closeAll()
+
+        self.assertEqual(closed, harness.windows)
+        self.assertEqual(harness.status_messages[-1][0], "Closing plot windows...")
+
+    def test_confirmation_options_use_shared_labels_and_config_keys(self):
+        updates = []
+
+        class FakeConfig:
+            values = {
+                "user_preference.confirm_close_all": True,
+                "user_preference.confirm_close": False,
+                }
+
+            def get(self, key):
+                return self.values[key]
+
+            def update(self, key, value):
+                updates.append((key, value))
+                self.values[key] = value
+
+        window = qtw.QMainWindow()
+        window.config = FakeConfig()
+        menu = qtw.QMenu(window)
+
+        try:
+            add_confirmation_options(window, menu)
+            actions = [
+                action for action in menu.actions()
+                if not action.isSeparator()
+                ]
+
+            self.assertEqual(
+                [action.text() for action in actions],
+                [
+                    "Confirm Before Closing All Plot Windows",
+                    "Confirm Before Quit",
+                    ]
+                )
+            self.assertTrue(actions[0].isChecked())
+            self.assertFalse(actions[1].isChecked())
+
+            actions[0].setChecked(False)
+            actions[1].setChecked(True)
+
+            self.assertEqual(updates, [
+                ("user_preference.confirm_close_all", False),
+                ("user_preference.confirm_close", True),
+                ])
+        finally:
+            window.deleteLater()
+
+    def test_restore_defaults_option_requests_main_window_reset(self):
+        called = []
+        window = qtw.QMainWindow()
+        window.restore_default_settings = lambda: called.append(True)
+        menu = qtw.QMenu(window)
+
+        try:
+            action = add_restore_defaults_option(window, menu)
+            self.assertEqual(action.text(), "Restore Default Settings...")
+
+            action.trigger()
+
+            self.assertEqual(called, [True])
+        finally:
+            window.deleteLater()
+
+    def test_restore_default_settings_can_be_cancelled(self):
+        old_question = qtw.QMessageBox.question
+
+        class FakeConfig:
+            def __init__(self):
+                self.reset_called = False
+
+            def reset_to_defaults(self):
+                self.reset_called = True
+
+        class Harness:
+            restore_default_settings = main_window.MainWindow.restore_default_settings
+
+            def __init__(self):
+                self.config = FakeConfig()
+                self.status_messages = []
+
+            def apply_current_settings(self):
+                raise AssertionError("Settings should not be applied after cancelling")
+
+            def show_status(self, message, timeout=5000):
+                self.status_messages.append((message, timeout))
+
+        try:
+            qtw.QMessageBox.question = lambda *args, **kwargs: qtw.QMessageBox.No
+            harness = Harness()
+            harness.restore_default_settings()
+        finally:
+            qtw.QMessageBox.question = old_question
+
+        self.assertFalse(harness.config.reset_called)
+        self.assertEqual(harness.status_messages[-1][0], "Default settings restore cancelled.")
+
+    def test_restore_default_settings_resets_and_applies_defaults(self):
+        old_question = qtw.QMessageBox.question
+
+        class FakeConfig:
+            def __init__(self):
+                self.reset_called = False
+
+            def reset_to_defaults(self):
+                self.reset_called = True
+
+        class Harness:
+            restore_default_settings = main_window.MainWindow.restore_default_settings
+
+            def __init__(self):
+                self.config = FakeConfig()
+                self.applied = False
+                self.status_messages = []
+
+            def apply_current_settings(self):
+                self.applied = True
+
+            def show_status(self, message, timeout=5000):
+                self.status_messages.append((message, timeout))
+
+        try:
+            qtw.QMessageBox.question = lambda *args, **kwargs: qtw.QMessageBox.Yes
+            harness = Harness()
+            harness.restore_default_settings()
+        finally:
+            qtw.QMessageBox.question = old_question
+
+        self.assertTrue(harness.config.reset_called)
+        self.assertTrue(harness.applied)
+        self.assertEqual(harness.status_messages[-1][0], "Default settings restored.")
+
+
+
+class DatabaseDropTestCase(unittest.TestCase):
+    def test_database_info_report_summarises_qcodes_tables(self):
+        with tempfile.TemporaryDirectory() as temp_dir:
+            database_path = str(Path(temp_dir) / "info.db")
+            conn = sqlite3.connect(database_path)
+            try:
+                conn.execute(
+                    "CREATE TABLE experiments (exp_id INTEGER PRIMARY KEY, name TEXT, sample_name TEXT)"
+                    )
+                conn.execute("""
+                  CREATE TABLE runs (
+                      run_id INTEGER PRIMARY KEY,
+                      name TEXT,
+                      run_timestamp REAL,
+                      completed_timestamp REAL,
+                      is_completed INTEGER,
+                      guid TEXT
+                  )
+                """)
+                conn.execute(
+                    "INSERT INTO experiments (exp_id, name, sample_name) VALUES (1, 'exp', 'sample')"
+                    )
+                conn.execute(
+                    """
+                    INSERT INTO runs
+                    (run_id, name, run_timestamp, completed_timestamp, is_completed, guid)
+                    VALUES (3, 'measurement', 1768129603, 1768129626, 1, 'guid-3')
+                    """
+                    )
+                conn.commit()
+            finally:
+                conn.close()
+
+            report = main_window.database_info_report(database_path)
+
+        self.assertIn("Runs: 1", report)
+        self.assertIn("Experiments: 1", report)
+        self.assertIn("Latest run ID: 3", report)
+        self.assertIn("Latest run GUID: guid-3", report)
+        self.assertIn("Database schema version:", report)
+        self.assertIn("Last modified:", report)
+        self.assertNotIn("Selected run ID:", report)
+        self.assertNotIn("Installed QCoDeS version:", report)
+        self.assertNotIn("QCoDeS active database:", report)
+        self.assertNotIn("SQLite version:", report)
+
+    def test_database_path_from_mime_data_accepts_one_local_db_file(self):
+        with tempfile.NamedTemporaryFile(suffix=".db") as database:
+            mime_data = QtCore.QMimeData()
+            mime_data.setUrls([QtCore.QUrl.fromLocalFile(database.name)])
+
+            self.assertEqual(
+                main_window.database_path_from_mime_data(mime_data),
+                database.name
+                )
+
+    def test_database_path_from_mime_data_rejects_ambiguous_or_non_db_drops(self):
+        with (
+            tempfile.NamedTemporaryFile(suffix=".db") as database,
+            tempfile.NamedTemporaryFile(suffix=".txt") as text_file,
+        ):
+            text_drop = QtCore.QMimeData()
+            text_drop.setUrls([QtCore.QUrl.fromLocalFile(text_file.name)])
+
+            multiple_drop = QtCore.QMimeData()
+            multiple_drop.setUrls([
+                QtCore.QUrl.fromLocalFile(database.name),
+                QtCore.QUrl.fromLocalFile(text_file.name),
+                ])
+
+            self.assertIsNone(main_window.database_path_from_mime_data(text_drop))
+            self.assertIsNone(main_window.database_path_from_mime_data(multiple_drop))
+
+
